Rename env to scope
